requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.20
langchain-core==0.1.20
google-generativeai==0.3.2
pymongo==4.6.0
streamlit==1.29.0
python-dotenv==1.0.0
PyPDF2==3.0.1
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
email-validator==2.1.0
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
plotly==5.19.0
faiss-cpu==1.7.4
google-cloud-aiplatform==1.38.1
langchain
langchain_community
langchain_openai

langchain_core